# docs/b2b/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BBB B2B Aggregator API",
    "version": "1.0.0",
    "description": "Verified MVP contract for the signed B2B API. All protected endpoints use HMAC headers documented in docs/api/HMAC_AUTHENTICATION.md."
  },
  "servers": [
    {
      "url": "{baseUrl}/api/b2b/v1",
      "variables": {
        "baseUrl": {
          "default": "https://operator-api.example"
        }
      }
    }
  ],
...
  "paths": {
...
    "/reports/settlements": {
      "get": {
        "tags": ["reports"],
        "summary": "List settlements for the signed operator",
        "description": "Returns a bounded tenant-scoped settlement list with validated filters and whitelisted sorting.",
        "security": [{ "B2BHmac": [] }],
        "parameters": [
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date" }, "example": "2026-06-01" },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date" }, "example": "2026-06-24" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 100 }, "example": 100 },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["draft", "exported", "submitted", "approved", "rejected"] }, "example": "exported" },
          { "name": "currency", "in": "query", "schema": { "type": "string", "minLength": 3, "maxLength": 3 }, "example": "USD" },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["created_at", "-created_at", "period_start", "-period_start", "period_end", "-period_end", "status", "-status", "currency", "-currency", "net_amount", "-net_amount", "settlement_uid", "-settlement_uid"], "default": "-created_at" }, "example": "-created_at" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "401": { "$ref": "#/components/responses/Error" },
          "422": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/reports/settlements/{settlement_uid}": {
      "get": {
        "tags": ["reports"],
        "summary": "Return one operator-scoped settlement snapshot and lifecycle state",
        "security": [{ "B2BHmac": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/SettlementUid" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "404": { "$ref": "#/components/responses/Error" },
          "422": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "B2BHmac": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Signature",
        "description": "Protected endpoints also require X-Operator-Id, X-Api-Key, X-Timestamp, X-Nonce, and X-Body-Hash. Signature payload is METHOD, path, canonical query, body hash, timestamp, nonce joined by newlines."
      }
    },
    "parameters": {
...
      "SettlementUid": {
        "name": "settlement_uid",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "maxLength": 80 },
        "example": "stl_abc123"
      }
    },
...
    "responses": {
      "Success": {
        "description": "Standard success envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiSuccess" }
          }
        }
      },
      "Error": {
        "description": "Standard error envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      }
    },
    "schemas": {
      "ApiSuccess": {
        "type": "object",
        "required": ["success", "status", "request_id", "data"],
        "properties": {
          "success": { "type": "boolean", "example": true },
          "status": { "type": "string", "example": "success" },
          "request_id": { "type": "string", "example": "uuid-or-client-request-id" },
          "data": { "type": "object", "additionalProperties": true },
          "meta": { "type": "object", "additionalProperties": true }
        }
      },
...
      "ApiError": {
        "type": "object",
        "required": ["success", "status", "request_id", "error"],
        "properties": {
          "success": { "type": "boolean", "example": false },
          "status": { "type": "string", "example": "error" },
          "request_id": { "type": "string", "example": "uuid-or-client-request-id" },
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "example": "VALIDATION_FAILED" },
              "message": { "type": "string", "example": "Request validation failed." },
              "details": { "type": "object", "additionalProperties": true }
            }
          },
          "meta": { "type": "object", "additionalProperties": true }
        }
      },
...
    }
  }
}